chore(@clayui/empty-state): move everything to index.tsx
